+++ src/main/java/net/geforcemods/securitycraft/SecurityCraft.java
@@ -87,6 +87,7 @@ public class SecurityCraft {
 	@SubscribeEvent
 	public static void onFMLCommonSetup(FMLCommonSetupEvent event) { //stage 1
 		RegistrationHandler.registerPackets();
+		RegistrationHandler.registerFakeLiquidRecipes();
 	}
 
 	@SubscribeEvent