+++ Changelog.md
@@ -6,6 +6,7 @@
 - New: Config option "codebreaker_chance" to set the chance of the codebreaker successfully hacking a block
 - Change: Codebreakers will now always open briefcases when in Creative mode without a chance of failing
 - Change: The Sentry's body is now solid if the sentry is not disguised
+- Change: The Fake Lava/Water Buckets are now created in a brewing stand, by putting the potion in the top slot, and the bucket(s) in the bottom slot(s)
 - API: Split IPasswordProtected into two interfaces (IPasswordProtected and ICodebreakable), meaning blocks can now be hacked by the codebreaker without needing to be password protected
 - API: Removed IPasswordProtected#isCodebreakable
 - Fix: The recipe for reinforced glass panes is incorrect